move util to the right place
